Create expandable_button_fan.dart
The closest I got to succeeding. It looks fine visually, but the hit test boxes don't move with the translations in AnimatedBuilder.

Potential issues:
Stack does not have a set size as it's not within a Container/Sizedbox. It is the immediate descendant of FractionalTranslation
Stack does not work well with Overlay technique.